mini_ecommerce_backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
dj-database-url==3.1.2
django-cors-headers==4.9.0
django-filter==25.2
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
idna==3.11
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
pyotp==2.9.0
python-dotenv==1.2.2
qrcode==8.2
reportlab==4.4.10
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
whitenoise==6.9.0
cloudinary==1.42.1
resend>=2.0.0
